small refactor

+++ src/kafka-net/Consumer.cs
@@ -154,10 +154,7 @@ namespace KafkaNet
 
                             var fetchRequest = new FetchRequest
                             {
-                                MaxWaitTime =
-                                    (int)
-                                        Math.Min((long)int.MaxValue,
-                                            _options.MaxWaitTimeForMinimumBytes.TotalMilliseconds),
+                                MaxWaitTime =(int)Math.Min(int.MaxValue,_options.MaxWaitTimeForMinimumBytes.TotalMilliseconds),
                                 MinBytes = _options.MinimumBytes,
                                 Fetches = fetches
                             };
